Crate bdk_tmp_plan

A spending plan or plan for short is a representation of a particular spending path on a descriptor. This allows us to analayze a choice of spending path without producing any signatures or other witness data for it.

To make a plan you provide the descriptor with “assets” like which keys you are able to use, hash pre-images you have access to, the current block height etc.

Once you’ve got a plan it can tell you its expected satisfaction weight which can be useful for doing coin selection. Furthermore it provides which subset of those keys and hash pre-images you will actually need as well as what locktime or sequence number you need to set.

Once you’ve obstained signatures, hash pre-images etc required by the plan, it can create a witness/script_sig for the input.

Structs

Assets

Plan

A plan represents a particular spending path for a descriptor.

PlanKey

A plan key contains the asset key originally provided along with key in the descriptor it purports to be able to derive for along with a “hint” on how to derive it.

Requirements

Signatures and hash pre-images that must be provided to complete the plan.

SatisfactionMaterial

Signatures and hash pre-images that can be used to complete a plan.

Enums

PlanState

The returned value from Plan::try_complete.

RequiredSignatures

The signatures required to complete the plan

SigningError

Traits

CanDerive

Functions

plan_satisfaction